% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CSV.R
\name{read_CSV}
\alias{read_CSV}
\title{Read Latest Version of CSV Saved With Initials and Date}
\usage{
read_CSV(filename, path = NULL, verbose = FALSE)
}
\arguments{
\item{filename}{String. The base name for the file.}

\item{path}{String. Where to look the file. If not set, checks for "Tables" or "_tables" in the working directory. Reads from working directory if data folders are not found.}

\item{verbose}{Logical. Should status messages be displayed?}
}
\description{
This function reads the latest version of an RDS file saved by save_CSV
}
\examples{
\dontrun{read_CSV(filename = "datafile")}
}
\keyword{CSV,}
\keyword{date}
\keyword{initials,}
\keyword{load,}
\keyword{read,}
