% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_RDS.R
\name{read_RDS}
\alias{read_RDS}
\title{Read Latest Version of RDS Saved With Initials and Date}
\usage{
read_RDS(filename, path = NULL, verbose = FALSE)
}
\arguments{
\item{filename}{String. The base name for the file.}

\item{path}{String. Where to look the file. If not set, checks for "Data" or "_data" in the working directory. Reads from working directory if data folders are not found.}

\item{verbose}{Logical. Should status messages be displayed?}
}
\description{
This function reads the latest version of an RDS file saved by save_RDS.
}
\examples{
\dontrun{read_RDS(filename = "datafile")}
}
\keyword{RDS,}
\keyword{date}
\keyword{initials,}
\keyword{load,}
\keyword{read,}
