% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_initials.R
\name{save_initials}
\alias{save_initials}
\title{Save Initials to .Renviron}
\usage{
save_initials(initials, overwrite = FALSE)
}
\arguments{
\item{initials}{String, containing users initials.}

\item{overwrite}{Logical. If initials are already in .Renvrion, should they be replaced?}
}
\description{
This function saves the user's initials to the .Renviron.
}
\examples{
\dontrun{save_initials("DA")}
}
\keyword{.Renviron}
\keyword{initials,}
\keyword{save,}
