% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_RDS.R
\name{save_RDS}
\alias{save_RDS}
\title{Save RDS With Initials and Date}
\usage{
save_RDS(
  object,
  filename,
  path = NULL,
  initials = Sys.getenv("UNLIKELY_INITIALS"),
  verbose = FALSE
)
}
\arguments{
\item{object}{The object in the R environment to save to a file.}

\item{filename}{String. The base name for the file.}

\item{path}{String. Where to save the file. If not set, checks for "Data" or "_data" in the working directory. Saves to working directory if data folders are not found.}

\item{initials}{String, or empty to retrieve from .Renviron.}

\item{verbose}{Logical. Should status messages be displayed?}
}
\description{
This function appends the user's initials and the current date when saving to RDS.
}
\examples{
save_initials("DA")
save_RDS(object = ls(), filename = "datafile")
}
\keyword{RDS,}
\keyword{date}
\keyword{initials,}
\keyword{save,}
