% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tex.R
\name{save_tex}
\alias{save_tex}
\title{Save LaTeX output to a renderable .tex document}
\usage{
save_tex(latex, file = NULL, landscape = FALSE)
}
\arguments{
\item{latex}{Character. The latex output from another function.}

\item{file}{Character. The (optional) file to save the output to.}

\item{landscape}{Logical. Should the document be saved as a PDF?}
}
\description{
This function takes LaTeX output from stargazer, texreg, xtable, etc. and saves it to an output file with the appropriate start and end notes.
}
\examples{
data("mtcars", envir = environment())
tex <- stargazer::stargazer(mtcars)
save_tex(tex)
}
