% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_wd.R
\name{set_wd}
\alias{set_wd}
\title{Set Cloud-Aware Working Directory}
\usage{
set_wd(project, cloud = NULL)
}
\arguments{
\item{project}{The name of the project directory. Must be set.}

\item{cloud}{The name of the cloud storage provider. If not set, the project is assumed to reside in your user folder.}
}
\description{
This function allows you to set a working directory within a given cloud storage folder in your home directory.
}
\examples{
set_wd("Platforms", "Box")
set_wd("Misogyny", "Dropbox")
set_wd("rcp-scraping")
}
\keyword{box,}
\keyword{cloud}
\keyword{directory,}
\keyword{dropbox,}
\keyword{odrive}
\keyword{storage,}
\keyword{working}
